Refactor to move FormSection under Forms
